Update libmagic to 5.47 and fix P/Invoke signatures
Fixes #62: magic_buffer returned application/octet-stream for ZIP files
due to a bug in libmagic 5.46's softmagic.c (upstream fix: file/file
commit 12b7664). Also fixes magic_file failing for ZIPs > 7MB
(MAGIC_PARAM_BYTES_MAX threshold).

- Rebuild linux-x64 libmagic-1.so and magic.mgc from file/file v5.47
- Fix P/Invoke size_t params (int -> nint) for magic_buffer,
  magic_getparam, and magic_setparam to prevent stack corruption
- Add MAGIC_NO_CHECK_SIMH flag (new in v5.47)
- Add ZIP detection tests for buffer, stream, and file path overloads

Note: other platform binaries (win-*, osx-*, linux-arm64, linux-musl-x64)
still need rebuilding from v5.47 source.

+++ src/Mime/MagicNative.cs
@@ -20,7 +20,7 @@ internal static class MagicNative
     public static extern IntPtr magic_file(IntPtr magic_cookie, string? dbPath);
 
     [DllImport(MAGIC_LIB_PATH, CallingConvention = CallingConvention.Cdecl)]
-    public static extern IntPtr magic_buffer(IntPtr magic_cookie, byte[] buffer, int length);
+    public static extern IntPtr magic_buffer(IntPtr magic_cookie, byte[] buffer, nint length);
 
     [DllImport(MAGIC_LIB_PATH, CallingConvention = CallingConvention.Cdecl)]
     public static extern IntPtr magic_error(IntPtr magic_cookie);
@@ -38,10 +38,10 @@ internal static class MagicNative
     public static extern int magic_compile(IntPtr magic_cookie, string? dbPath);
 
     [DllImport(MAGIC_LIB_PATH, CallingConvention = CallingConvention.Cdecl)]
-    public static extern int magic_getparam(IntPtr magic_cookie, MagicParams param, out int value);
+    public static extern int magic_getparam(IntPtr magic_cookie, MagicParams param, out nint value);
 
     [DllImport(MAGIC_LIB_PATH, CallingConvention = CallingConvention.Cdecl)]
-    public static extern int magic_setparam(IntPtr magic_cookie, MagicParams param, ref int value);
+    public static extern int magic_setparam(IntPtr magic_cookie, MagicParams param, ref nint value);
 
     [DllImport(MAGIC_LIB_PATH, CallingConvention = CallingConvention.Cdecl)]
     public static extern int magic_version();
